Refactor: Move observability tests to test_observability.py
